pyproject.toml
[project]
name = "nerve-source-gmail"
version = "0.1.0"
description = "Gmail source adapter for Insilica Nerve"
requires-python = ">=3.11"
dependencies = [
    "insilica-nerve",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[project.scripts]
nerve-source-gmail = "nerve_source_gmail.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nerve_source_gmail"]
